% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\name{animals}
\alias{animals}
\title{Aniamls Function}
\usage{
animals(shark, cat, cow)
}
\arguments{
\item{shark}{shark}

\item{cat}{cat}

\item{cow}{cow}
}
\value{
number
}
\description{
Aniamls Function
}
\examples{
animals(shark = 1, cat = 2, cow = 4)
}
