% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grocery_bill.R
\name{grocery_bill}
\alias{grocery_bill}
\title{Grocery Bill}
\usage{
grocery_bill(tacos, banana, sando)
}
\arguments{
\item{tacos}{# number of tacos}

\item{banana}{#number of banana}

\item{sando}{#number of sandos}
}
\value{
# cost of grocery bill
}
\description{
Grocery Bill
}
\examples{
grocery_bill(tacos, banana, sando)
}
