% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallory.R
\name{mallory}
\alias{mallory}
\title{Mallory}
\usage{
mallory(person, outfit, time_of_day)
}
\arguments{
\item{person}{#state person}

\item{outfit}{#state outfit}

\item{time_of_day}{#state time of day}
}
\value{
# stuff
}
\description{
Mallory
}
\examples{
mallory(lewis, shoes, afternoon)
}
