% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snacks.R
\name{snacks}
\alias{snacks}
\title{Calculate my Snacks}
\usage{
snacks(grapes, apples, burritos, oranges)
}
\arguments{
\item{grapes}{# number of grapes}

\item{apples}{# number of apples}

\item{burritos}{# number of burritos}

\item{oranges}{#nunber of oranges}
}
\value{
# calculation
}
\description{
Calculate my Snacks
}
\examples{
snacks(grapes = 2.5, apples = 6, burritos = 1, oranges = 2)
}
